Unblock full CI collection for dev PR
Keep full-suite pytest collection and architecture boundaries stable before merging the dev integration branch.

Constraint: CI runs pytest over the entire tests tree on both Ubuntu and Windows, and pytest imports duplicate test basenames as the same top-level module.

Rejected: Add onboarding -> cli to the architecture allowlist | The dependency direction is wrong; onboarding should not depend on CLI presentation modules.

Confidence: high

Scope-risk: narrow

Directive: Keep shared terminal UI helpers outside the cli package when onboarding or gateway surfaces need them.

Tested: uv run --no-sync pytest tests/test_ci/test_architecture_import_contracts.py tests/test_onboarding/test_flow.py::test_readme_distinguishes_recommended_profile_from_channel_extras tests/test_channels/test_channel_manager_lifecycle.py -q

Tested: uv run --no-sync ruff check src tests

Tested: uv run --no-sync mypy src/opensquilla --show-error-codes

Tested: uv run --no-sync pytest tests -q

Tested: uv run --no-sync python -m pytest tests/test_gateway/test_task_runtime_terminal_cleanup.py::test_no_leak_under_load -v

Tested: uv build --wheel

Not-tested: GitHub hosted CI rerun after push.
