+++ tests/test_channels/test_channel_manager_lifecycle.py
